+++ integration/crud-typeorm/companies/companies.controller.ts
@@ -4,14 +4,12 @@ import { Crud } from '@nestjsx/crud';
 
 import { Company } from './company.entity';
 import { CompaniesService } from './companies.service';
-import { dto } from './dto';
-import { serialize } from './response';
+import { serialize } from './responses';
 
 @Crud({
   model: {
     type: Company,
   },
-  // dto,
   serialize,
   routes: {
     deleteOneBase: {
@@ -19,10 +17,27 @@ import { serialize } from './response';
     },
   },
   query: {
-    alwaysPaginate: true,
+    alwaysPaginate: false,
+    allow: ['name'],
     join: {
-      users: {},
-      projects: {},
+      users: {
+        alias: 'companyUsers',
+        exclude: ['email'],
+        eager: true,
+      },
+      'users.projects': {
+        eager: true,
+        alias: 'usersProjects',
+        allow: ['name'],
+      },
+      'users.projects.company': {
+        eager: true,
+        alias: 'usersProjectsCompany',
+      },
+      projects: {
+        eager: true,
+        select: false,
+      },
     },
   },
 })