+++ integration/crud-typeorm/notes/note.entity.ts
@@ -1,0 +1,10 @@
+import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity('notes')
+export class Note {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column({ name: 'revision_id', nullable: false })
+  revisionId: number;
+}