+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [4.6.1] - 2020-05-08
+
+### Bug Fixes
+
+- **typeorm** - fixed query generation when a column display name differs from its name in db [#401](https://github.com/nestjsx/crud/issues/401)). Kudos to @farhad2161 for finding this bug
+
+## [4.6.0] - 2020-05-07
+
+### Features
+
+- **crud**/**typeorm** - added `select` (boolean) to `join` options which allows to join relation but not select it ([#218](https://github.com/nestjsx/crud/issues/218))
+
+### Bug Fixes
+
+- **typeorm** - fixed column identifier for MySQL ([#401](https://github.com/nestjsx/crud/issues/401))
+- **typeorm** - fixed nested relations aliases, filtering, sorting ([#419](https://github.com/nestjsx/crud/issues/419), [#450](https://github.com/nestjsx/crud/issues/450), [#267](https://github.com/nestjsx/crud/issues/267), [#385](https://github.com/nestjsx/crud/issues/385))
+
 ## [4.5.0] - 2020-05-01
 
 ### Improvements
@@ -148,6 +165,8 @@
 
 - several fixes
 
+[4.6.1]: https://github.com/nestjsx/crud/compare/v4.6.0...v4.6.1
+[4.6.0]: https://github.com/nestjsx/crud/compare/v4.5.0...v4.6.0
 [4.5.0]: https://github.com/nestjsx/crud/compare/v4.4.5...v4.5.0
 [4.4.5]: https://github.com/nestjsx/crud/compare/v4.4.4...v4.4.5
 [4.4.4]: https://github.com/nestjsx/crud/compare/v4.4.3...v4.4.4